backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2

# Real-time Communication
python-socketio==5.10.0
python-engineio==4.7.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Notifications
twilio==8.10.0
firebase-admin==6.4.0
sendgrid==6.10.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1